annotated plots, move nullspace to utils
