+++ src/mud/plot.py
@@ -1,70 +1,11 @@
 from matplotlib import pyplot as plt
 import numpy as np
+from mud.util import null_space
 
-def null_space(A, rcond=None):
-    """
-    Construct an orthonormal basis for the null space of A using SVD
-
-    Parameters
-    ----------
-    A : (M, N) array_like
-        Input array
-    rcond : float, optional
-        Relative condition number. Singular values ``s`` smaller than
-        ``rcond * max(s)`` are considered zero.
-        Default: floating point eps * max(M,N).
-
-    Returns
-    -------
-    Z : (N, K) ndarray
-        Orthonormal basis for the null space of A.
-        K = dimension of effective null space, as determined by rcond
-
-    See also
-    --------
-    svd : Singular value decomposition of a matrix
-    orth : Matrix range
-
-    Examples
-    --------
-    One-dimensional null space:
-
-    >>> from scipy.linalg import null_space
-    >>> A = np.array([[1, 1], [1, 1]])
-    >>> ns = null_space(A)
-    >>> ns * np.sign(ns[0,0])  # Remove the sign ambiguity of the vector
-    array([[ 0.70710678],
-           [-0.70710678]])
-
-    Two-dimensional null space:
-
-    >>> B = np.random.rand(3, 5)
-    >>> Z = null_space(B)
-    >>> Z.shape
-    (5, 2)
-    >>> np.allclose(B.dot(Z), 0)
-    True
-
-    The basis vectors are orthonormal (up to rounding error):
-
-    >>> Z.T.dot(Z)
-    array([[  1.00000000e+00,   6.92087741e-17],
-           [  6.92087741e-17,   1.00000000e+00]])
-
-    """
-    u, s, vh = np.linalg.svd(A, full_matrices=True)
-    M, N = u.shape[0], vh.shape[1]
-    if rcond is None:
-        rcond = np.finfo(s.dtype).eps * max(M, N)
-    tol = np.amax(s) * rcond
-    num = np.sum(s > tol, dtype=int)
-    Q = vh[num:,:].T.conj()
-    return Q
-
-def plotChain(mud_chain, ref_param, color='k'):
+def plotChain(mud_chain, ref_param, color='k', s=100):
     num_steps = len(mud_chain)
     current_point = mud_chain[0]
-    plt.scatter(current_point[0], current_point[1], c='b')
+    plt.scatter(current_point[0], current_point[1], c='b', s=s)
     for i in range(0,num_steps):
         next_point = mud_chain[i]
         points = np.hstack([current_point, next_point])
@@ -74,16 +15,17 @@ def plotChain(mud_chain, ref_param, color='k'):
     plt.ylim([0,1])
     plt.xlim([0,1])
 #     plt.axis('off')
-    plt.scatter(ref_param[0], ref_param[1], c='r')
+    plt.scatter(ref_param[0], ref_param[1], c='r', s=s)
 
 
-def plot_contours(A, ref_param, color='k', ls=':', lw=1, **kwds):
+def plot_contours(A, ref_param, color='k', ls=':', lw=1, fs=20, w=1, s=100, **kwds):
     numQoI = A.shape[0]
     AA = np.hstack([null_space(A[i,:].reshape(1,-1)) for i in range(numQoI)]).T
     for i in range(numQoI):
-        plt.plot([ref_param[0]-AA[i,0], ref_param[1]+AA[i,0]],
-                 [ref_param[0]-AA[i,1], ref_param[1]+AA[i,1]],
-                 c=color, ls=ls, lw=lw, **kwds)
+        xloc = [ref_param[0] - w*AA[i,0], ref_param[1] + w*AA[i,0]]
+        yloc = [ref_param[0] - w*AA[i,1], ref_param[1] + w*AA[i,1]]
+        plt.plot(xloc, yloc, c=color, ls=ls, lw=lw, **kwds)
+        plt.annotate('%d'%i, (xloc[0], yloc[0]), fontsize=fs)
 
 
 def make_2d_normal_mesh(N=50, window=1):
